Refine file structure.
