+++ src/assets/manifest.json
@@ -4,7 +4,7 @@
   "version": "1.0",
   "manifest_version": 3,
   "permissions": ["activeTab", "scripting", "bookmarks", "storage", "tabs"],
-  "host_permissions": ["*://*/*", "chrome://*/*"],
+  "host_permissions": ["*://*/*"],
   "background": {
     "service_worker": "background.js"
   },
